Moved tests for game searching to `tests/Feature/Game`
